fix time step width computation
In the previous version, the time step size fitting (+10% or refine the time step, or dt = inteval_length / n_steps) was already applied during construction when the time interval is not known.
Therefore it was applied to the default [0,1] interval which resulted in wrong time step sizes. This lead to errors in the systemMatrix.
Especially when setting a fixed number of time steps per interval, the results were very wrong.

We now store the step size given in the settings file and recompute the time step correctly every time the integration interval changes.

A follow up commit will fix the systemMatrix setup, to use the correct time step size.